Moved physics sim to its own workspace
