+++ physics/Cargo.toml
@@ -4,3 +4,6 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+bevy = { version = "0.14.1", features = ["dynamic_linking"] }
+bevy_infinite_grid = "0.13.0"
+log = { version = "*", features = ["max_level_debug", "release_max_level_warn"] } # Remove trace logging at compile time
